+++ cypress/integration/visitorCanSeeFullRecipe.feature.js
@@ -50,27 +50,12 @@ describe("A visitor, by clicking a recipe card in the main view", () => {
       "February 07, 2022 16:38"
     );
   });
-
   it("is expected to display a add comment field", () => {
     cy.get("[data-cy=comment-field]").should("be.visible");
   });
 
   it("is expected to display a comment feed", () => {
     cy.get("[data-cy=comment-feed]").should("be.visible");
-  });
-
-  describe("post a comment when hitting the enter button", () => {
-    before(() => {
-      cy.get("[data-cy=comment-field]")
-        .type("I really enjoyed this recipe!")
-        .type("{enter}");
-    });
-    it("is expected to display a comment in the comment feed", () => {
-      cy.get("[data-cy=comment-feed").should(
-        "contain.text",
-        "I really enjoyed this recipe!"
-      );
-    });
   });
 
   describe("goes back to main page", () => {
@@ -88,5 +73,27 @@ describe("A visitor, by clicking a recipe card in the main view", () => {
     it("is expected to see a collection of recipes", () => {
       cy.get("[data-cy=recipes-list]").children().should("have.length", 7);
     });
+  });
+});
+
+describe.only("post a comment when hitting the enter button", () => {
+  before(() => {
+    cy.intercept("GET", "/api/recipes/**", {
+      fixture: "recipesShowResponse.json"
+    }).as("RecipeShow");
+    cy.visit("/recipes/12");
+
+    cy.intercept("POST", "/api/recipes/**/comments", {
+      fixture: "recipesShowResponse.json"
+    });
+    cy.get("[data-cy=comment-field]")
+      .type("I really enjoyed this recipe!")
+      .type("{enter}");
+  });
+  it("is expected to display a comment in the comment feed", () => {
+    cy.get("[data-cy=comment-feed").should(
+      "contain.text",
+      "I really enjoyed this recipe!"
+    );
   });
 });
