+++ src/modules/Recipes.js
@@ -59,25 +59,7 @@ const Recipes = {
       return error;
     }
   },
-  async create(id) {
-    try {
-      const headers = JSON.parse(localStorage.getItem("J-tockAuth-Storage"));
-      const response = await api.post(
-        `/recipes/${id}/comments`,
-        {
-          comment: {
-            body: "I really liked your focaccia"
-          }
-        },
-        {
-          headers: headers
-        }
-      );
-      return response;
-    } catch (error) {
-      return error;
-    }
-  },
+  
   async update(recipe) {
     try {
       const headers = JSON.parse(localStorage.getItem("J-tockAuth-Storage"));
